Base.c		I added a BaseLayerReparent method that takes care of 		changing the wkptr in the BasePart of the object.  It 		also calls _NhlReparent on all the children of the 		object, so their wkptr can be updated as well.
BaseP.h		I changed the function prototype for the BaseLayerReparent
		method because more information needed to be passed.

Converters.c	Changed a cpp check for "sun" to check for "SunOs".

CoordArrTable.c	Modified it so that individual vectors in the data can
		be NULL, they are assumed to contain thier integer index.
		I also made the diType resource "gettable".

DataComm.c	Fixed a bug in the "init" procedure.  It was looking at
		the wrong "layer_class" record.

Draw.c		I added "doc" tags.

GetValues.c	""

Reparent.c	Implimented the BaseLayerReparent method.  Before it didn't
		really call the method in the class - it just traversed
		the instance tree itself.  Now it is implimented threw the
		method.

SetValues.c	Added "doc" tags.

TickMark.c	Removed extraneous casts - they aren't portable.

XyPlot.c	Added support for the NULL vectors in the CoordArrTable
		object.

hluP.h		Added a prototype for the new private function _NhlReparent.

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1735 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/hlu/TickMark.c
@@ -1,5 +1,5 @@
 /*
- *      $Id: TickMark.c,v 1.5 1993-12-13 23:34:46 ethan Exp $
+ *      $Id: TickMark.c,v 1.6 1994-01-10 19:48:58 boote Exp $
  */
 /************************************************************************
 *									*
@@ -2340,12 +2340,12 @@ static float	compare
 */
 	if(a_final>b_final){ 
 		tmp = (float)log10(a_final);
-		exp = (long)(float)ceil(log10(a_final));
+		exp = (long)ceil(log10(a_final));
 		if((float)exp == tmp)
 			exp++;
 	} else {
 		tmp = (float)log10(b_final);
-		exp = (long)(float)ceil(log10(b_final));
+		exp = (long)ceil(log10(b_final));
 		if((float)exp == tmp)
 			exp++;
 	}
